[fix] 전략패턴 적용에 따른 Car 클래스 추가 및 전략 설정 클래스 추가, 랜덤레이스 카 클래스 삭제-> 랜덤이동 전략으로 변경
