+++ src/main/java/study/step3/Car.java
@@ -1,6 +1,19 @@
 package study.step3;
-public interface Car{
-    void move(boolean condition);
-    public int getPosition();
-    public String getCarName();
+
+public class Car extends Moving{
+    private static final Integer START_POSITION = 1;
+    private String carName;
+    private int position;
+    public Car(String carName){
+        this.carName = carName;
+        this.position = START_POSITION;
+    }
+
+    public int getPosition(){
+        return position;
+    }
+    public String getCarName(){
+        return carName;
+    }
+
 }
